# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ticket-comments": {
            "get": {
                "description": "Get all comments, optionally filtered by ticket ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-comments"
                ],
                "summary": "Get all ticket comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by Ticket ID",
                        "name": "ticket_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.TicketComment"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new comment on a ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-comments"
                ],
                "summary": "Create a new ticket comment",
                "parameters": [
                    {
                        "description": "Comment Data",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateTicketCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketCommentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        },
        "/ticket-comments/{id}": {
            "get": {
                "description": "Get a specific comment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-comments"
                ],
                "summary": "Get a ticket comment by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TicketComment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-comments"
                ],
                "summary": "Update a ticket comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Comment Data",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateTicketCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketCommentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing comment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ticket-comments"
                ],
                "summary": "Delete a ticket comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        },
...
        "/tickets": {
            "get": {
                "description": "Get a list of all tickets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get all tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/requests.TicketResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Create a new ticket",
                "parameters": [
                    {
                        "description": "Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Get a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Get a ticket by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update a ticket by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Update a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Ticket Data",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.TicketCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/requests.TicketResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a ticket by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Delete a ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "validation": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "assignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAssignment"
                    }
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAttachment"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.TicketCategory"
                },
                "category_id": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketComment"
                    }
                },
                "deskripsi": {
                    "type": "string"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "kode_tiket": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketLog"
                    }
                },
                "priority": {
                    "$ref": "#/definitions/models.TicketPriority"
                },
                "priority_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.TicketStatus"
                },
                "status_id": {
                    "type": "integer"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "tanggal_diperbarui": {
                    "type": "string"
                },
                "tipe_pengaduan": {
                    "type": "string"
                },
                "user": {
                    "description": "Relasi - Add references to match custom column names",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.TicketAssignment": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/models.User"
                },
                "id_admin": {
                    "type": "integer"
                },
                "id_assignment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "tanggal_ditugaskan": {
                    "type": "string"
                },
                "ticket": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                }
            }
        },
        "models.TicketAttachment": {
            "type": "object",
            "properties": {
                "file_path": {
                    "type": "string"
                },
                "id_attachment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "ticket": {
                    "description": "Relasi - specify references",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                },
                "uploaded_at": {
                    "type": "string"
                }
            }
        },
        "models.TicketCategory": {
            "type": "object",
            "properties": {
                "id_category": {
                    "type": "integer"
                },
                "nama_category": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                }
            }
        },
        "models.TicketComment": {
            "type": "object",
            "properties": {
                "id_comment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "isi_pesan": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "ticket": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    ]
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.TicketPriority": {
            "type": "object",
            "properties": {
                "id_priority": {
                    "type": "integer"
                },
                "nama_priority": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "admin_availability": {
                    "$ref": "#/definitions/models.AdminAvailability"
                },
                "admin_states": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AdminConversationState"
                    }
                },
                "assigned_tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketAssignment"
                    }
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketComment"
                    }
                },
                "conversations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Conversation"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketLog"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ticket"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "requests.CreateTicketCommentRequest": {
            "type": "object",
            "properties": {
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "isi_pesan": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                }
            }
        },
...
        "requests.TicketCommentResponse": {
            "type": "object",
            "properties": {
                "id_comment": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "isi_pesan": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                }
            }
        },
        "requests.TicketCreateRequest": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string",
                    "example": "Saya tidak bisa login ke akun saya."
                },
                "id_category": {
                    "type": "integer",
                    "example": 1
                },
                "id_priority": {
                    "type": "integer",
                    "example": 1
                },
                "id_status": {
                    "type": "integer",
                    "example": 1
                },
                "id_user": {
                    "type": "integer",
                    "example": 1
                },
                "judul": {
                    "type": "string",
                    "example": "Tidak bisa login"
                },
                "kode_tiket": {
                    "type": "string",
                    "example": "TCKT-001"
                },
                "tipe_pengaduan": {
                    "description": "Allowed: pelanggan, penjual",
                    "type": "string",
                    "enum": [
                        "pelanggan",
                        "penjual"
                    ],
                    "example": "pelanggan"
                }
            }
        },
        "requests.TicketResponse": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "id_category": {
                    "type": "integer"
                },
                "id_priority": {
                    "type": "integer"
                },
                "id_status": {
                    "type": "integer"
                },
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "judul": {
                    "type": "string"
                },
                "kode_tiket": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                },
                "tanggal_diperbarui": {
                    "type": "string"
                },
                "tipe_pengaduan": {
                    "type": "string"
                }
            }
        },
        "requests.UpdateTicketCommentRequest": {
            "type": "object",
            "properties": {
                "id_ticket": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "isi_pesan": {
                    "type": "string"
                },
                "tanggal_dibuat": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}